# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/totp/v1/register" : {
      "post" : {
        "summary" : "处理注册请求，生成并返回URI和QR码。",
        "description" : "处理注册请求，生成并返回URI和QR码。",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "username" ],
                "type" : "object",
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "description" : "用户名"
                  }
                },
                "description" : "注册请求对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "注册响应对象，包含URI和QR码",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "uri" : {
                      "type" : "string",
                      "description" : "URI"
                    },
                    "qr" : {
                      "type" : "string",
                      "description" : "QR码"
                    }
                  },
                  "description" : "注册响应对象，包含URI和QR码"
                }
              }
            }
          }
        }
      }
    },
    "/totp/v1/verify" : {
      "post" : {
        "summary" : "处理验证请求，验证TOTP是否有效。",
        "description" : "处理验证请求，验证TOTP是否有效。",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "totp", "username" ],
                "type" : "object",
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "description" : "用户名"
                  },
                  "totp" : {
                    "type" : "integer",
                    "description" : "TOTP"
                  }
                },
                "description" : "验证请求对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "验证响应对象，包含验证结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "valid" : {
                      "type" : "boolean",
                      "description" : "验证是否有效"
                    }
                  },
                  "description" : "验证响应对象，包含验证结果"
                }
              }
            }
          }
        }
      }
    },
    "/otp/v1/send" : {
      "post" : {
        "summary" : "发送OTP请求。",
        "description" : "发送OTP请求。",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "destination", "sender" ],
                "type" : "object",
                "properties" : {
                  "destination" : {
                    "type" : "string",
                    "description" : "目的地号码，不能为空"
                  },
                  "sender" : {
                    "type" : "string",
                    "description" : "发送者号码，不能为空"
                  }
                },
                "description" : "发送OTP请求对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "发送OTP的结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "sessionId" : {
                      "type" : "string",
                      "description" : "会话ID"
                    },
                    "destination" : {
                      "type" : "string",
                      "description" : "目的地"
                    },
                    "resendAllowedAt" : {
                      "type" : "integer",
                      "description" : "允许重发时间",
                      "format" : "int64"
                    },
                    "remainingAttempts" : {
                      "type" : "integer",
                      "description" : "剩余尝试次数"
                    }
                  },
                  "description" : "发送OTP的结果"
                }
              }
            }
          }
        }
      }
    },
    "/otp/v1/verify" : {
      "post" : {
        "summary" : "验证OTP请求。",
        "description" : "验证OTP请求。",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "otp", "sessionId" ],
                "type" : "object",
                "properties" : {
                  "sessionId" : {
                    "type" : "string",
                    "description" : "会话ID"
                  },
                  "otp" : {
                    "type" : "string",
                    "description" : "OTP验证码"
                  }
                },
                "description" : "验证OTP请求对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "验证OTP的结果",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "valid" : {
                      "type" : "boolean",
                      "description" : "验证是否有效"
                    },
                    "remainingAttempts" : {
                      "type" : "integer",
                      "description" : "剩余尝试次数"
                    }
                  },
                  "description" : "验证OTP的结果"
                }
              }
            }
          }
        }
      }
    }
  },
...
}